Fixed rails revision number for checkout. They break too much stuff. Switched to new_<blah> for nearly everything in GroupsController. Nuked all login/password code. YAY! I succumbed to anti-YAGNI thinking. Changed all has_many to have exclusively_dependent marked true. Stupid. Added address to Location. Added google_url to Location live during Seattle.rb meeting. Hooked google_url into the group show page. Fixed the database config. Extended seattle.rb script to work as a total reset, or just reset seattle only. Fixed all broken test cases from yipstar's work. Surprisingly easy actually!
[git-p4: depot-paths = "//src/rubyholic/dev/": change = 2259]

+++ script/seattle.rb
@@ -1,14 +1,21 @@
 #!/usr/local/bin/ruby
 
-ENV["RAILS_ENV"] = 'development'
+ENV["RAILS_ENV"] ||= 'development'
 require File.dirname(__FILE__) + '/../config/environment'
 
-Group.delete_all
+if ARGV.empty? then
+  Group.delete_all
 
-seattle = Group.new
-seattle.name = "Seattle Ruby Brigade!"
-seattle.city = "Seattle, WA"
-seattle.save
+  seattle = Group.new
+  seattle.name = "Seattle Ruby Brigade!"
+  seattle.city = "Seattle, WA"
+  seattle.save
+else
+  seattle = Group.find_by_city "Seattle, WA"
+  seattle.urls.clear 
+  seattle.contacts.clear
+  seattle.locations.clear
+end
 
 seattle.new_url "web", "http://www.zenspider.com/seattle.rb"
 seattle.new_url "mail", "mailto:ruby@zenspider.com"
@@ -17,13 +24,21 @@ seattle.new_url "irc", "irc://irc.freenode.net/#seattle.rb"
 ryan = seattle.new_contact "Ryan Davis", "ryand-ruby@zenspider.com"
 eric = seattle.new_contact "Eric Hodel", "drbrain@segment7.net"
 
-amazon = seattle.new_location "Amazon US1"
-robot  = seattle.new_location "Robot Co-op"
+amazon = seattle.new_location "Amazon US1", "605 5th Ave, Seattle, WA 98104"
+robot  = seattle.new_location "Robot Co-op", "1205 E Pike #2F, Seattle, WA 98102"
+redline  = seattle.new_location "Red Line Cafe", "1525 E Olive Way, Seattle, WA 98122"
 
 monthly = seattle.new_event "Monthly Meeting", "2005-11-29 19:00", amazon
-weekly  = seattle.new_event "Weekly Meeting", "2005-11-15 19:00", robot
+monthly.new_subject "Joe Heitzeberg on Asterisk and R/AGI"
 
+monthly = seattle.new_event "Monthly Meeting", "2005-12-27 19:00", redline
+monthly.new_subject "Eric Hodel on Hard Core TDD for Rails"
 monthly.new_subject "Evan Webb on Sydney"
-monthly.new_subject "Eric Hodel on TDD for Rails"
 
-
+seattle.new_event "Weekly Hacking", "2005-11-01 19:00", robot
+seattle.new_event "Weekly Hacking", "2005-11-08 19:00", robot
+seattle.new_event "Weekly Hacking", "2005-11-15 19:00", robot
+seattle.new_event "Weekly Hacking", "2005-11-22 19:00", robot
+seattle.new_event "Weekly Hacking", "2005-12-06 19:00", redline
+seattle.new_event "Weekly Hacking", "2005-12-13 19:00", redline
+seattle.new_event "Weekly Hacking", "2005-12-20 19:00", redline
